requirements.txt
gradio
requests
beautifulsoup4
python-dotenv
playwright